backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
pymongo==4.7.2
groq==0.11.0
httpx==0.27.2
geopy==2.4.1
APScheduler==3.10.4
python-dotenv==1.0.1
